feat(agent): add External Services section and move Exa API key
- Merge AgentSettingsDialog and AgentToolsDialog into a single settings dialog
- Add CollapsibleSection grouping with 'Provider settings', 'External services', and 'Tools' sections
- Move Exa API key field from provider settings into the new External Services section
- Remove standalone AgentToolsDialog component and AgentHeader Tools button
- Use shared CollapsibleSection component for consistent 4px group spacing
- Add exaApiKey to provider settings store and copy strings
- Update agent-workspace test indices for new field order
