+++ packages/vue/src/features/agent/agent-copy.ts
@@ -7,7 +7,7 @@ const copy = {
     newSession: 'New session',
     renameSession: 'Rename session',
     deleteSession: 'Delete session',
-    settings: 'Provider settings',
+    settings: 'Agent settings',
     closePanel: 'Close Agent panel',
     switchLanguage: 'Switch language',
     connected: 'Connected',
@@ -53,6 +53,9 @@ const copy = {
     usage: 'Usage',
     tokens: 'tokens',
     providerTitle: 'Model provider',
+    agentSettings: 'Agent settings',
+    providerSettings: 'Provider settings',
+    externalServices: 'External services',
     tools: 'Tools',
     noTools: 'No Agent tools are registered.',
     toolParameters: 'Parameters (JSON)',
@@ -108,7 +111,7 @@ const copy = {
     newSession: '新建会话',
     renameSession: '重命名会话',
     deleteSession: '删除会话',
-    settings: 'Provider 设置',
+    settings: 'Agent 设置',
     closePanel: '关闭 Agent 面板',
     switchLanguage: '切换语言',
     connected: '已连接',
@@ -154,6 +157,9 @@ const copy = {
     usage: '用量',
     tokens: 'tokens',
     providerTitle: '模型 Provider',
+    agentSettings: 'Agent 设置',
+    providerSettings: '供应商设置',
+    externalServices: '外部服务',
     tools: '工具',
     noTools: '当前没有已注册的 Agent 工具。',
     toolParameters: '参数（JSON）',
