+++ packages/vue/src/features/agent/components/AgentWorkspace.vue
@@ -11,7 +11,6 @@
       @rename="renameSession"
       @delete="deleteSession"
       @settings="providerSettings.show(state.provider)"
-      @tools="providerSettings.showTools()"
       @close="$emit('close')"
       @toggle-locale="toggleLocale"
     />
@@ -52,7 +51,6 @@
       :status="state.provider"
       :locale="locale"
     />
-    <AgentToolsDialog :provider-settings="providerSettings" :locale="locale" />
   </section>
 </template>
 
@@ -67,7 +65,6 @@
   import AgentHeader from './AgentHeader.vue'
   import AgentSettingsDialog from './AgentSettingsDialog.vue'
   import AgentTimeline from './AgentTimeline.vue'
-  import AgentToolsDialog from './AgentToolsDialog.vue'
 
   import type { AgentBridgeClient } from '../agent-contracts'
